Refactor site initialization templates
