+++ docs/api/init/scaffolding.md
@@ -1,0 +1,8 @@
+# Initialization - Scaffolding
+
+MkDocs site structure creation and management utilities.
+
+::: egregora.init.scaffolding
+    options:
+      show_source: true
+      show_root_heading: true
