--- docs/api/publication/scaffolding.md
@@ -1,8 +1,0 @@
-# Publication - Scaffolding
-
-MkDocs site structure creation and management.
-
-::: egregora.publication.site.scaffolding
-    options:
-      show_source: true
-      show_root_heading: true
