+++ src/egregora/templates/site/README.md.jinja2
