+++ src/egregora/templates/site/docs/about.md.jinja2
