+++ src/egregora/templates/site/docs/media/index.md.jinja2
@@ -1,0 +1,11 @@
+# Media Library
+
+Este diretório organiza todos os anexos processados pela Egregora.
+
+- 🖼️ **images/** — fotos e capturas de tela compartilhadas no grupo
+- 🎬 **videos/** — trechos de vídeo enviados pelos participantes
+- 🎧 **audio/** — mensagens de voz e gravações
+- 📄 **documents/** — PDFs, planilhas e outros arquivos relevantes
+
+Arquivos são salvos automaticamente durante o processamento e servem como
+referência para os posts publicados.
