+++ src/egregora/templates/site/docs/posts/index.md.jinja2
