+++ src/egregora/templates/site/mkdocs.yml.jinja2
